# shared/openapi.yaml
openapi: 3.1.0
info:
  title: cinghialapp API
  version: 0.1.0
  description: |
    D&D 5e campaign manager API.
    Server-authoritative dice. Role-based visibility (player/master).
servers:
  - url: http://localhost:8080/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /campaigns/{id}/factions: { parameters: [{ $ref: '#/components/parameters/Id' }], get: { tags: [world], responses: { '200': { description: list } } }, post: { tags: [world], responses: { '201': { description: created } } } }
  /campaigns/{id}/npcs:     { parameters: [{ $ref: '#/components/parameters/Id' }], get: { tags: [world], responses: { '200': { description: list } } }, post: { tags: [world], responses: { '201': { description: created } } } }
  /campaigns/{id}/lore:     { parameters: [{ $ref: '#/components/parameters/Id' }], get: { tags: [world], responses: { '200': { description: list } } }, post: { tags: [world], responses: { '201': { description: created } } } }
  /campaigns/{id}/news:     { parameters: [{ $ref: '#/components/parameters/Id' }], get: { tags: [world], responses: { '200': { description: list } } }, post: { tags: [world], responses: { '201': { description: created } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...